+++ source/renderer/assets/codemirror/zettlr-plugin-render-math.js
@@ -55,7 +55,7 @@
     let fromLine = i
 
     for (let i = 0; i < cm.lineCount(); i++) {
-      if (cm.getModeAt({ 'line': i, 'ch': 0 }).name !== 'markdown-zkn') continue
+      if (cm.getModeAt({ 'line': i, 'ch': 0 }).name !== 'markdown') continue
       // Reset the index of the expression everytime we enter a new line.
       inlineMathRE.lastIndex = 0